Move vimiv.gui.widgets to vimiv.widgets
The widgets are not only used by the gui and importing from the gui in
other (utility) modules is confusing as the gui is the outermost layer.
